Ability to change title and description + updates server
also made edit values better; previously had a different state for each and
different onChange function. made them all part of the same state now, kind of
like credentials is in eisenhour.

also fixed a little bug introduced in the last commit; in handleSubmit when
elapsed is recalculated, i last time forgot to change the dates passed to it
from startTime and endTime to newStartTime and newEndTime (the mutated clones)

also added {/* comments */} to the jsx