fix(sync): Bug that was causing UnhandledPromiseRejectionWarning ..
.. upon server error.

Upon server error, the insert/update/delete functions reject. However I did not
have a return there before, so it was continuing to execute the code after the
rejection, which in some places was accessing res.data which is not set in the
case of error, causing further error and unhandled promise rejection.